// src/common/enums/twilio-provisioning-status.enum.ts
export enum TwilioProvisioningStatus {
  CREATING_SUBACCOUNT = 'CREATING_SUBACCOUNT',
  PURCHASING_NUMBER = 'PURCHASING_NUMBER',
  CONFIGURING_VOICE = 'CONFIGURING_VOICE',
  ACTIVATING = 'ACTIVATING',
  ACTIVE = 'ACTIVE',
  FAILED = 'FAILED',
  SUSPENDED = 'SUSPENDED',
  CLOSING = 'CLOSING',
  CLOSED = 'CLOSED',
}